MLPXOR
An MLP therefore, known as a deep artificial neural network. It is a collection of more than one perceptron. They are composed of an input layer to receive the signal, an output layer that makes a choice or prediction about the input, and in between, an random no. of hidden layers (True computational engine of the MLP).